Refs #127. Merged the StylesManager into the StyleDefinitionPart
